src/main/java/Artwork.java
import java.util.ArrayList;

public class Artwork {
    private String title;
    private String artist;
    private double price;
    private int nft;


    public Artwork(String title, String artist, double price, int nft) {
        this.title = title;
        this.artist = artist;
        this.price = price;
        this.nft = nft;
    }

    public String getTitle() {
        return this.title;
    }

    public String getArtist() {
        return this.artist;
    }

    public double getPrice() {
        return this.price;}

    public int getNft() {
            return this.nft;}


    public int buyArtworkFromGallery() {
        return 0;
    }


    public void setArtworks(int i) {
    }
}
